コンポーネントを分離
